LP64 sweep: fix pointer truncation, an out-of-bounds write, and a UAF
A deliberate pass for the same 32-to-64-bit class that produced the invisible
torches, rather than waiting for the next one to surface as a crash. Five real
defects, all invisible on Win32 where long is 4 bytes.

  ds1.c bounds-checked every read of the loaded DS1 against the end of the
  buffer as (UDWORD) ptr < ((UDWORD) ds1_buff) + ds1_len. UDWORD is uint32_t,
  so both addresses were truncated to their low 32 bits before comparing, and
  the sum could wrap on top of that. Whether the check answered correctly
  depended on where malloc placed the buffer, and it could fail either way:
  refuse a pointer that was fine, leaving group and NPC-path fields silently
  zeroed, or accept one that was not and read off the end of the allocation.
  Six sites, now comparing pointers as pointers.

  ds1.c read the workspace save_count with sscanf("%lu") into a UDWORD field.
  That stores an unsigned long -- eight bytes into a four-byte field, so it
  wrote four bytes past it into whatever the struct puts there. The matching
  sprintf("%lu") on the write side read an eight-byte vararg for the same
  four-byte value, so the count written to the file had four bytes of
  whatever followed it in the frame.

  compose_render.c decoded DC6 frame headers through a `long *` that was
  assigned an int32_t cast, so every field was read eight bytes wide at twice
  its offset -- fh[1] picked up fields 2 and 3 instead of the width. The frame
  data pointer had the same fault from the other end: 8 * sizeof(long) is 64
  here and the header is 32 bytes. This is the fallback for bosses whose
  layers ship as DC6, and it produced nothing at all: exporting Mephisto left
  an empty directory before, and gives a correct 155x171 sprite after. dc6.c
  reads the identical layout correctly and is what this now matches.

  txtread.c built its "calloc() failed" message by dereferencing the struct it
  had just passed to txt_destroy, which frees it and returns NULL. The error
  path crashed instead of reporting.

  66 printf/sprintf format specifiers disagreed with their arguments -- %li
  against a 32-bit value or %i against a long. Undefined in varargs and able
  to desync the rest of the call. Mostly diagnostics, except the save_count
  one above.

Checked and found clean, so the next person need not redo it: every packed
on-disk struct (only ANIM_DATA_RECORD_S, already fixed width), the base
typedefs in types.h, INDEX_COLORMAP, every fwrite/fread size against its
field, and the remaining 0x7FFFFFFF/0x80000000 constants (all correct -- ints,
bitmasks, or negated). Wav_unp.c accounts for 396 int-to-pointer-cast warnings
but is already compiled out on 64-bit behind MPQ_WAV_UNP_USABLE.

Left alone deliberately: txtread stores numeric columns as long while cof.c
reads them as int32_t. Little-endian makes that correct for every value these
tables hold, and changing the column width touches the hot path of all table
parsing, so it is documented rather than churned.

Warnings 198 -> 117 excluding Wav_unp.c, 0 errors, 28/28 tests, --headless and
--selftest both render.

+++ src/core/dcc.c
@@ -52,7 +52,7 @@ int dcc_read_bytes(DCC_BITSTREAM_S * bs, int bytes_number, void * dest)
       {
          sprintf(dcc_error, "dcc_read_bytes() : want to read behind "
                             "the end (%li >= %li)\n",
-                            bs->cur_bit_num, bs->size);
+                            (long)bs->cur_bit_num, (long)bs->size);
          return 1;
       }
       * (((UBYTE *) dest) + i) = bs->data[bs->cur_byte];
@@ -113,7 +113,7 @@ int dcc_read_bits(DCC_BITSTREAM_S * bs, int bits_number, int is_signed,
       {
          sprintf(dcc_error, "dcc_read_bits() : want to read behind "
                             "the end (%li >= %li)\n",
-                            bs->cur_bit_num, bs->size);
+                            (long)bs->cur_bit_num, (long)bs->size);
          return 1;
       }
 
@@ -402,7 +402,7 @@ int dcc_optional_datas(DCC_S * dcc, DCC_BITSTREAM_S * bs, int d)
                "dcc_optional_datas() : can't allocate "
                "%li bytes for optional_byte_data\n"
                "direction %i, frame %i\n",
-               size, d, f);
+               (long)size, d, f);
             return 1;
          }
 
@@ -778,7 +778,7 @@ int dcc_fill_pixel_buffer(DCC_S * dcc, int d)
    {
       sprintf(dcc_error, "dcc_fill_pixel_buffer() : can't create buffer bitmap "
                          "of %i * %i pixels\n",
-                         dir->box.width, dir->box.height);
+                         (int)dir->box.width, (int)dir->box.height);
       return 1;
    }
    a5_clear(dir->bmp); // all 0
@@ -1088,7 +1088,7 @@ int dcc_make_frames(DCC_S * dcc, int d)
    {
       sprintf(dcc_error, "dcc_make_frames() : can't create tmp frame "
                          "bitmap of %i * %i pixels",
-                         dir->box.width, dir->box.height
+                         (int)dir->box.width, (int)dir->box.height
       );
       return 1;
    }
@@ -1521,15 +1521,15 @@ void dcc_debug(DCC_S * dcc)
    for (d=0; d<dcc->header.directions; d++)
    {
       printf("\ndirection %2i\n", d);
-      printf("   outsize_coded       = %li bytes\n", dcc->direction[d].outsize_coded);
-      printf("   compression_flag    = %li\n", dcc->direction[d].compression_flag);
-      printf("   variable0_bits      = %2lu  (%2i bits)\n", dcc->direction[d].variable0_bits,      dcc_bits_width_table[dcc->direction[d].variable0_bits]);
-      printf("   width_bits          = %2lu  (%2i bits)\n", dcc->direction[d].width_bits,          dcc_bits_width_table[dcc->direction[d].width_bits]);
-      printf("   height_bits         = %2lu  (%2i bits)\n", dcc->direction[d].height_bits,         dcc_bits_width_table[dcc->direction[d].height_bits]);
-      printf("   xoffset_bits        = %2lu  (%2i bits)\n", dcc->direction[d].xoffset_bits,        dcc_bits_width_table[dcc->direction[d].xoffset_bits]);
-      printf("   yoffset_bits        = %2lu  (%2i bits)\n", dcc->direction[d].yoffset_bits,        dcc_bits_width_table[dcc->direction[d].yoffset_bits]);
-      printf("   optional_bytes_bits = %2lu  (%2i bits)\n", dcc->direction[d].optional_bytes_bits, dcc_bits_width_table[dcc->direction[d].optional_bytes_bits]);
-      printf("   coded_bytes_bits    = %2lu  (%2i bits)\n", dcc->direction[d].coded_bytes_bits,    dcc_bits_width_table[dcc->direction[d].coded_bytes_bits]);
+      printf("   outsize_coded       = %li bytes\n", (long)dcc->direction[d].outsize_coded);
+      printf("   compression_flag    = %li\n", (long)dcc->direction[d].compression_flag);
+      printf("   variable0_bits      = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].variable0_bits,      dcc_bits_width_table[dcc->direction[d].variable0_bits]);
+      printf("   width_bits          = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].width_bits,          dcc_bits_width_table[dcc->direction[d].width_bits]);
+      printf("   height_bits         = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].height_bits,         dcc_bits_width_table[dcc->direction[d].height_bits]);
+      printf("   xoffset_bits        = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].xoffset_bits,        dcc_bits_width_table[dcc->direction[d].xoffset_bits]);
+      printf("   yoffset_bits        = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].yoffset_bits,        dcc_bits_width_table[dcc->direction[d].yoffset_bits]);
+      printf("   optional_bytes_bits = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].optional_bytes_bits, dcc_bits_width_table[dcc->direction[d].optional_bytes_bits]);
+      printf("   coded_bytes_bits    = %2lu  (%2i bits)\n", (unsigned long)dcc->direction[d].coded_bytes_bits,    dcc_bits_width_table[dcc->direction[d].coded_bytes_bits]);
       printf("   box                 = (%li, %li)  --->  (%li, %li)  =  %li * %li\n",
          dcc->direction[d].box.xmin,
          dcc->direction[d].box.ymin,
@@ -1547,11 +1547,11 @@ void dcc_debug(DCC_S * dcc)
          dcc->direction[d].box.xmin,
          dcc->direction[d].box.ymax
       );
-      printf("   equal_cell_bitstream_size       = %8li bits (read %8li)\n", dcc->direction[d].equal_cell_bitstream_size, dcc->direction[d].equal_cell_bitstream.cur_bit_num);
-      printf("   pixel_mask_bitstream_size       = %8li bits (read %8li)\n", dcc->direction[d].pixel_mask_bitstream_size, dcc->direction[d].pixel_mask_bitstream.cur_bit_num);
-      printf("   encoding_type_bitstream_size    = %8li bits (read %8li)\n", dcc->direction[d].encoding_type_bitstream_size, dcc->direction[d].encoding_type_bitstream.cur_bit_num);
-      printf("   raw_pixel_bitstream_size        = %8li bits (read %8li)\n", dcc->direction[d].raw_pixel_bitstream_size, dcc->direction[d].raw_pixel_bitstream.cur_bit_num);
-      printf("   pixel_code_and_displacment.size = %8li bits (read %8li)\n", dcc->direction[d].pixel_code_and_displacment_bitstream.size, dcc->direction[d].pixel_code_and_displacment_bitstream.cur_bit_num);
+      printf("   equal_cell_bitstream_size       = %8li bits (read %8li)\n", (long)dcc->direction[d].equal_cell_bitstream_size, (long)dcc->direction[d].equal_cell_bitstream.cur_bit_num);
+      printf("   pixel_mask_bitstream_size       = %8li bits (read %8li)\n", (long)dcc->direction[d].pixel_mask_bitstream_size, (long)dcc->direction[d].pixel_mask_bitstream.cur_bit_num);
+      printf("   encoding_type_bitstream_size    = %8li bits (read %8li)\n", (long)dcc->direction[d].encoding_type_bitstream_size, (long)dcc->direction[d].encoding_type_bitstream.cur_bit_num);
+      printf("   raw_pixel_bitstream_size        = %8li bits (read %8li)\n", (long)dcc->direction[d].raw_pixel_bitstream_size, (long)dcc->direction[d].raw_pixel_bitstream.cur_bit_num);
+      printf("   pixel_code_and_displacment.size = %8li bits (read %8li)\n", (long)dcc->direction[d].pixel_code_and_displacment_bitstream.size, (long)dcc->direction[d].pixel_code_and_displacment_bitstream.cur_bit_num);
 
       printf("\n   pixel_values :\n");
       for (i=0; i < 256; i++)
@@ -1566,13 +1566,13 @@ void dcc_debug(DCC_S * dcc)
          printf("\n   frame %3i\n", f);
 
          if (dcc->direction[d].variable0_bits)
-            printf("      variable0      = %lu\n", dcc->frame[d][f].variable0);
+            printf("      variable0      = %lu\n", (unsigned long)dcc->frame[d][f].variable0);
 
          if (dcc->direction[d].width_bits)
-            printf("      width          = %lu\n", dcc->frame[d][f].width);
+            printf("      width          = %lu\n", (unsigned long)dcc->frame[d][f].width);
 
          if (dcc->direction[d].height_bits)
-            printf("      height         = %lu\n", dcc->frame[d][f].height);
+            printf("      height         = %lu\n", (unsigned long)dcc->frame[d][f].height);
 
          if (dcc->direction[d].xoffset_bits)
             printf("      xoffset        = %li\n", dcc->frame[d][f].xoffset);
@@ -1581,12 +1581,12 @@ void dcc_debug(DCC_S * dcc)
             printf("      yoffset        = %li\n", dcc->frame[d][f].yoffset);
 
          if (dcc->direction[d].optional_bytes_bits)
-            printf("      optional_bytes = %lu\n", dcc->frame[d][f].optional_bytes);
+            printf("      optional_bytes = %lu\n", (unsigned long)dcc->frame[d][f].optional_bytes);
 
          if (dcc->direction[d].coded_bytes_bits)
-            printf("      coded_bytes    = %lu\n", dcc->frame[d][f].coded_bytes);
+            printf("      coded_bytes    = %lu\n", (unsigned long)dcc->frame[d][f].coded_bytes);
 
-         printf("      bottom_up      = %li\n", dcc->frame[d][f].bottom_up);
+         printf("      bottom_up      = %li\n", (long)dcc->frame[d][f].bottom_up);
          printf("      box            = (%li, %li)  --->  (%li, %li)  =  %li * %li\n",
             dcc->frame[d][f].box.xmin,
             dcc->frame[d][f].box.ymin,
@@ -1612,8 +1612,8 @@ void dcc_debug(DCC_S * dcc)
          );
 
          printf("      1st cell0_x, y = %li, %li\n",
-            dcc->frame[d][f].cell[0].x0,
-            dcc->frame[d][f].cell[0].y0
+            (long)dcc->frame[d][f].cell[0].x0,
+            (long)dcc->frame[d][f].cell[0].y0
          );
 
          if (dcc->frame[d][f].optional_bytes)
